+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/laf/DefaultLookAndFeelProvider.java
@@ -25,10 +25,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Consumer;
 
-import static is.codion.swing.common.ui.Utilities.systemLookAndFeelClassName;
-import static is.codion.swing.common.ui.laf.LookAndFeelProvider.lookAndFeelProvider;
 import static java.util.Objects.requireNonNull;
-import static javax.swing.UIManager.getCrossPlatformLookAndFeelClassName;
 
 final class DefaultLookAndFeelProvider implements LookAndFeelProvider {
 
@@ -37,16 +34,6 @@ final class DefaultLookAndFeelProvider implements LookAndFeelProvider {
 	static final Map<String, LookAndFeelProvider> LOOK_AND_FEEL_PROVIDERS = new HashMap<>();
 
 	static {
-		if (CROSS_PLATFORM.getOrThrow()) {
-			LookAndFeelProvider crossPlatformProvider = lookAndFeelProvider(new LookAndFeelInfo("Cross Platform", getCrossPlatformLookAndFeelClassName()));
-			LOOK_AND_FEEL_PROVIDERS.put(crossPlatformProvider.lookAndFeelInfo().getClassName(), crossPlatformProvider);
-		}
-		if (SYSTEM.getOrThrow()) {
-			LookAndFeelProvider systemProvider = lookAndFeelProvider(new LookAndFeelInfo("System", systemLookAndFeelClassName()));
-			if (!LOOK_AND_FEEL_PROVIDERS.containsKey(systemProvider.lookAndFeelInfo().getClassName())) {
-				LOOK_AND_FEEL_PROVIDERS.put(systemProvider.lookAndFeelInfo().getClassName(), systemProvider);
-			}
-		}
 		LookAndFeelProviders.instances().forEach(providers ->
 						providers.get().forEach(LookAndFeelProviders::addLookAndFeel));
 	}