+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/laf/package-info.java
@@ -19,9 +19,7 @@
 /**
  * Package configuration values:
  * <ul>
- * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelProvider#CROSS_PLATFORM}
- * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelProvider#SYSTEM}
- * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelComboBox#ENABLE_ON_SELECTION}
+ * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelComboBox#DEFAULT_LOOK_AND_FEEL}
  * </ul>
  */
 package is.codion.swing.common.ui.laf;